Move ProxyToolbar model to its own module
