+++ bokehjs/src/lib/models/tools/index.ts
@@ -35,5 +35,5 @@ export {Tool}              from "./tool"
 export {ToolProxy}         from "./tool_proxy"
 export {Toolbar}           from "./toolbar"
 export {ToolbarBase}       from "./toolbar_base"
-export {ProxyToolbar}      from "./toolbar_box"
+export {ProxyToolbar}      from "./proxy_toolbar"
 export {ToolbarBox}        from "./toolbar_box"
